LPS-123758 Move classes that are not exported to internal package - frontend-js
