+++ modules/apps/frontend-js/frontend-js-minifier/src/test/java/com/liferay/frontend/js/minifier/internal/GoogleJavascriptMinifierTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.frontend.js.minifier;
+package com.liferay.frontend.js.minifier.internal;
 
 import com.liferay.portal.kernel.test.CaptureHandler;
 import com.liferay.portal.kernel.test.JDKLoggerTestUtil;
